+++ dbt_platform_helper/providers/config.py
@@ -1,4 +1,5 @@
 import os
+from copy import deepcopy
 from pathlib import Path
 
 import click
@@ -10,7 +11,6 @@ from yamllint import linter
 
 from dbt_platform_helper.constants import PLATFORM_CONFIG_FILE
 from dbt_platform_helper.providers.platform_config_schema import PlatformConfigSchema
-from dbt_platform_helper.utils.files import apply_environment_defaults
 from dbt_platform_helper.utils.messages import abort_with_error
 
 
@@ -44,7 +44,7 @@ class ConfigProvider:
         PlatformConfigSchema.schema().validate(self.config)
 
         # TODO= logically this isn't validation but loading + parsing, to move.
-        enriched_config = apply_environment_defaults(self.config)
+        enriched_config = self.apply_environment_defaults(self.config)
         self.validator.run_validations(enriched_config)
 
     def load_and_validate_platform_config(
@@ -78,3 +78,37 @@ class ConfigProvider:
                 "Please check it exists and you are in the root directory of your deployment project."
             )
             exit(1)
+
+    def apply_environment_defaults(self, config):
+        if "environments" not in config:
+            return config
+
+        enriched_config = deepcopy(config)
+
+        environments = enriched_config["environments"]
+        env_defaults = environments.get("*", {})
+        without_defaults_entry = {
+            name: data if data else {} for name, data in environments.items() if name != "*"
+        }
+
+        default_versions = config.get("default_versions", {})
+
+        def combine_env_data(data):
+            return {
+                **env_defaults,
+                **data,
+                "versions": {
+                    **default_versions,
+                    **env_defaults.get("versions", {}),
+                    **data.get("versions", {}),
+                },
+            }
+
+        defaulted_envs = {
+            env_name: combine_env_data(env_data)
+            for env_name, env_data in without_defaults_entry.items()
+        }
+
+        enriched_config["environments"] = defaulted_envs
+
+        return enriched_config
