+++ tests/platform_helper/providers/test_config.py
@@ -1,4 +1,5 @@
 from pathlib import Path
+from unittest.mock import Mock
 from unittest.mock import patch
 
 import pytest
@@ -343,3 +344,144 @@ def test_load_and_validate_platform_config_skips_file_check_when_disable_file_ch
         config_provider.load_and_validate_platform_config(disable_file_check=True)
 
         assert not mock_config_file_check.called
+
+
+def test_apply_defaults():
+    config_provider = ConfigProvider(Mock())
+
+    config = {
+        "application": "my-app",
+        "environments": {
+            "*": {"a": "aaa", "b": {"c": "ccc"}},
+            "one": None,
+            "two": {},
+            "three": {"a": "override_aaa", "b": {"d": "ddd"}, "c": "ccc"},
+        },
+    }
+
+    result = config_provider.apply_environment_defaults(config)
+
+    assert result == {
+        "application": "my-app",
+        "environments": {
+            "one": {"a": "aaa", "b": {"c": "ccc"}, "versions": {}},
+            "two": {"a": "aaa", "b": {"c": "ccc"}, "versions": {}},
+            "three": {"a": "override_aaa", "b": {"d": "ddd"}, "c": "ccc", "versions": {}},
+        },
+    }
+
+
+@pytest.mark.parametrize(
+    "default_versions, env_default_versions, env_versions, expected_result",
+    [
+        # Empty cases
+        (None, None, None, {}),
+        (None, None, {}, {}),
+        (None, {}, None, {}),
+        ({}, None, None, {}),
+        # Env versions populated
+        (
+            None,
+            None,
+            {"platform-helper": "8.0.0", "terraform-platform-modules": "1.0.0"},
+            {"platform-helper": "8.0.0", "terraform-platform-modules": "1.0.0"},
+        ),
+        (
+            None,
+            None,
+            {"terraform-platform-modules": "2.0.0"},
+            {"terraform-platform-modules": "2.0.0"},
+        ),
+        (None, None, {"platform-helper": "9.0.0"}, {"platform-helper": "9.0.0"}),
+        # env_default_versions populated
+        (None, {"platform-helper": "10.0.0"}, None, {"platform-helper": "10.0.0"}),
+        (None, {"platform-helper": "10.0.0"}, {}, {"platform-helper": "10.0.0"}),
+        (
+            None,
+            {"platform-helper": "10.0.0"},
+            {"platform-helper": "8.0.0", "terraform-platform-modules": "1.0.0"},
+            {"platform-helper": "8.0.0", "terraform-platform-modules": "1.0.0"},
+        ),
+        (
+            None,
+            {"platform-helper": "10.0.0"},
+            {"terraform-platform-modules": "2.0.0"},
+            {"platform-helper": "10.0.0", "terraform-platform-modules": "2.0.0"},
+        ),
+        # default_versions populated
+        (
+            None,
+            {"platform-helper": "10.0.0"},
+            {"platform-helper": "9.0.0"},
+            {"platform-helper": "9.0.0"},
+        ),
+        (
+            {"terraform-platform-modules": "1.0.0"},
+            None,
+            None,
+            {"terraform-platform-modules": "1.0.0"},
+        ),
+        (
+            {"terraform-platform-modules": "2.0.0"},
+            {"terraform-platform-modules": "3.0.0"},
+            None,
+            {"terraform-platform-modules": "3.0.0"},
+        ),
+        (
+            {"terraform-platform-modules": "3.0.0"},
+            None,
+            {"terraform-platform-modules": "4.0.0"},
+            {"terraform-platform-modules": "4.0.0"},
+        ),
+        (
+            {"terraform-platform-modules": "4.0.0"},
+            {"terraform-platform-modules": "5.0.0"},
+            {"terraform-platform-modules": "6.0.0"},
+            {"terraform-platform-modules": "6.0.0"},
+        ),
+    ],
+)
+def test_apply_defaults_for_versions(
+    default_versions, env_default_versions, env_versions, expected_result
+):
+    config = {
+        "application": "my-app",
+        "environments": {"*": {}, "one": {}},
+    }
+    config_provider = ConfigProvider(Mock())
+
+    if default_versions:
+        config["default_versions"] = default_versions
+    if env_default_versions:
+        config["environments"]["*"]["versions"] = env_default_versions
+    if env_versions:
+        config["environments"]["one"]["versions"] = env_versions
+
+    result = config_provider.apply_environment_defaults(config)
+
+    assert result["environments"]["one"].get("versions") == expected_result
+
+
+def test_apply_defaults_with_no_defaults():
+    config = {
+        "application": "my-app",
+        "environments": {
+            "one": None,
+            "two": {},
+            "three": {
+                "a": "aaa",
+            },
+        },
+    }
+    config_provider = ConfigProvider(Mock())
+
+    result = config_provider.apply_environment_defaults(config)
+
+    assert result == {
+        "application": "my-app",
+        "environments": {
+            "one": {"versions": {}},
+            "two": {"versions": {}},
+            "three": {"a": "aaa", "versions": {}},
+        },
+    }
